Support for assessments and assessment entries
Data structures are not great, but it's functional. Minimal-ish example
included: `assessment_details`.

In particular, AssessmentDetails.raw_data should be "correct" because it
is scraped straight from the tables. AssessmentDetails.data is hashed
based on the column ID and multiple attempts at an assessment MAY make
weird things happen there.